requirements.txt
# Torch/torchaudio installed via Dockerfile (CPU index) before this file
whisperx>=3.8.1
fastapi>=0.133.0
uvicorn[standard]>=0.41.0
python-multipart>=0.0.22
structlog>=25.5.0

# OpenTelemetry (traces + metrics), optional — enable via OTEL_ENABLED=true
opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0
opentelemetry-exporter-otlp-proto-http>=1.28.0
opentelemetry-instrumentation-fastapi>=0.49b0
